+++ frontend/app/routes/$lang+/_protected+/personal-information+/home-address+/address-accuracy.tsx
@@ -3,17 +3,22 @@ import { json } from '@remix-run/node';
 import { Form, useLoaderData } from '@remix-run/react';
 
 import { useTranslation } from 'react-i18next';
+import { redirectWithSuccess } from 'remix-toast';
 
 import pageIds from '../../../page-ids.json';
 import { Address } from '~/components/address';
 import { Button, ButtonLink } from '~/components/buttons';
+import { getAddressService } from '~/services/address-service.server';
+import { getAuditService } from '~/services/audit-service.server';
 import { getInstrumentationService } from '~/services/instrumentation-service.server';
 import { getLookupService } from '~/services/lookup-service.server';
 import { getPersonalInformationService } from '~/services/personal-information-service.server';
 import { getRaoidcService } from '~/services/raoidc-service.server';
+import { getUserService } from '~/services/user-service.server';
 import { getTypedI18nNamespaces } from '~/utils/locale-utils';
-import { getFixedT, redirectWithLocale } from '~/utils/locale-utils.server';
+import { getFixedT, getLocale, redirectWithLocale } from '~/utils/locale-utils.server';
 import { mergeMeta } from '~/utils/meta-utils';
+import { IdToken } from '~/utils/raoidc-utils.server';
 import type { UserinfoToken } from '~/utils/raoidc-utils.server';
 import { getTitleMetaTags } from '~/utils/seo-utils';
 
@@ -60,13 +65,27 @@ export async function loader({ context: { session }, request }: LoaderFunctionAr
 }
 
 export async function action({ context: { session }, request }: ActionFunctionArgs) {
+  const addressService = getAddressService();
   const instrumentationService = getInstrumentationService();
   const raoidcService = await getRaoidcService();
+  const userService = getUserService();
 
   await raoidcService.handleSessionValidation(request, session);
 
+  const userId = await userService.getUserId();
+  const userInfo = await userService.getUserInfo(userId);
+
+  const newHomeAddress = session.get('useSuggestedAddress') ? session.get('suggestedAddress') : session.get('newHomeAddress');
+  await addressService.updateAddressInfo(userId, userInfo?.homeAddress ?? '', newHomeAddress);
+
+  const idToken: IdToken = session.get('idToken');
+  getAuditService().audit('update-data.home-address', { userId: idToken.sub });
+
+  const locale = getLocale(request);
+
   instrumentationService.countHttpStatus('home-address.validate', 302);
-  return redirectWithLocale(request, '/personal-information/home-address/confirm');
+  // TODO remove new home address from session and handle case when it is missing
+  return redirectWithSuccess(`/${locale}/personal-information`, 'personal-information:home-address.address-accuracy.updated-notification');
 }
 
 export default function PersonalInformationHomeAddressAccuracy() {
