+++ frontend/public/locales/fr/personal-information.json
@@ -39,7 +39,8 @@
       "button": {
         "change": "(FR) Change",
         "cancel": "(FR) Cancel"
-      }
+      },
+      "updated-notification": "(FR) Home address updated"
     },
     "address-accuracy": {
       "page-title": "(FR) Home Address Accuracy",
@@ -52,7 +53,8 @@
       "invalid-address-info": "(FR) Your address is compared against Canada Post information for accuracy. However, in some cases, the system may not be able to verify the address's accuracy. Use this address if you are certain it is correct.",
       "note": "(FR) Please note: Choose \"Continue\" if the home address is correct or choose \"Re-enter address\" if changes are required",
       "requested-change": "(FR) Requested change:",
-      "re-enter-address": "(FR) Re-enter your address"
+      "re-enter-address": "(FR) Re-enter your address",
+      "updated-notification": "(FR) Home address updated"
     },
     "suggested": {
       "page-title": "(FR) Suggested address",
@@ -70,21 +72,7 @@
       "note": "(FR) Please note: Your address is compared against Canada Post information for accuracy. However, in some cases, the system may not be able to verify the address's accuracy. Use this address if you are certain it is correct.",
       "continue": "(FR) Continue",
       "cancel": "(FR) Cancel",
-      "edit": "(FR) Return to edit page"
-    },
-    "confirm": {
-      "page-title": "(FR) Confirm",
-      "subtitle": "(FR) Please confirm the home address change below.",
-      "breadcrumbs": {
-        "personal-information": "(FR) Personal information",
-        "address-change-confirm": "(FR) Confirm"
-      },
-      "previous-home-address": "(FR) Previous home address",
-      "new-home-address": "(FR) New home address",
-      "button": {
-        "confirm": "(FR) Confirm",
-        "cancel": "(FR) Cancel"
-      },
+      "edit": "(FR) Return to edit page",
       "updated-notification": "(FR) Home address updated"
     }
   },
